requirements.txt
praw
pyyaml
pytest
